+++ rapo/algorithms/reconciliation/s08_save_error_a.sql
@@ -13,7 +13,7 @@ select {parallelism} *
            from rapo_temp_source_a_{process_id} a
                 left join rapo_temp_t01_mod_{process_id} m on {key_field_a} = m.a_id and m.correlation_indicator is not null
                 left join rapo_temp_t02_org_a_{process_id} o on {key_field_a} = o.a_id
-          where {date_field_a} between to_date('{date_from:%Y-%m-%d %H:%M:%S}', 'YYYY-MM-DD HH24:MI:SS') and to_date('{date_to:%Y-%m-%d %H:%M:%S}', 'YYYY-MM-DD HH24:MI:SS')
+          where {window_a}
             and (o.correlation_type in ('O', 'F') or o.correlation_type is null)
             and (
                   ((o.correlation_type = 'O' or o.correlation_type is null) and (o.correlation_indicator is null or (o.correlation_indicator is not null and (m.discrepancy_time_a is not null {discrepancy_filters_a})))) or
@@ -32,7 +32,7 @@ select {parallelism} *
                 case when m.discrepancy_time_a is not null then m.discrepancy_time_a||'|'||m.discrepancy_time_value||';' end{discrepancy_descriptions_a} as rapo_discrepancy_description
            from rapo_temp_source_a_{process_id} a left join rapo_temp_t02_org_a_{process_id} o on {key_field_a} = o.a_id
                 left join rapo_temp_t01_mod_{process_id} m on {key_field_a} = m.a_id and m.distance_rank_a = 1
-          where {date_field_a} between to_date('{date_from:%Y-%m-%d %H:%M:%S}', 'YYYY-MM-DD HH24:MI:SS') and to_date('{date_to:%Y-%m-%d %H:%M:%S}', 'YYYY-MM-DD HH24:MI:SS')
+          where {window_a}
             and (o.correlation_type in ('A', 'B', 'M') or (o.correlation_type = 'F' and o.correlation_indicator is null))
        )
  where rapo_result_type in ({target_error_types_a})